+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.1.23] - 2022-05-23
+### Fixed
+- Properly display range of more than 2 months on Datepicker
+- Fix ref prop on input components
+
 ## [2.1.22] - 2022-05-11
 ### Changed
 - Improve color typing prop
@@ -333,6 +338,7 @@
 ### Changed
 - Updated gap and styles on Row component
 
+[2.1.23]: https://github.com/marshmallow-insurance/smores-react/compare/v2.1.21...v2.1.23
 [2.1.22]: https://github.com/marshmallow-insurance/smores-react/compare/v2.1.21...v2.1.22
 [2.1.21]: https://github.com/marshmallow-insurance/smores-react/compare/v2.1.20...v2.1.21
 [2.1.20]: https://github.com/marshmallow-insurance/smores-react/compare/v2.1.19...v2.1.20